requirements.txt
adafruit-io==2.4.0
awscrt==0.5.15
awsiotsdk==1.1.0
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
paho-mqtt==1.5.0
py-sds011==0.9
pyserial==3.4
python-aqi==0.6.1
requests==2.23.0
urllib3==1.25.9
python-dotenv==0.13.0